blog token tests

+++ controllers/blog.js
@@ -61,7 +61,19 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id)
+	const token = getTokenFrom(request)
+	const decodedToken = jwt.verify(token, process.env.SECRET)
+	if (!token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const user = await User.findById(decodedToken.id)
+	const blog = await Blog.findById(request.params.id)
+
+	if ( blog.user.toString() === user._id.toString() ){
+		await Blog.findByIdAndRemove(request.params.id)
+	}else{
+		return response.status(401).json({ error: 'unauthorized user' })
+	}
 	response.status(200).end()
 })
